requirements.txt
webrtcvad==2.0.10
google-cloud-speech==2.26.0
